Mashall Command type from Dhall
I followed the docs from here:

https://hackage.haskell.org/package/dhall-1.30.0/docs/Dhall.html#t:FromDhall

and looked at the `FromDhall` instance for `Fix f`. I fixed the Haskell
compilation errors that I got, and all of a sudden the Haskell code
compiled.

When I ran Laps, I got error messages from Dhall mentioning the schema
that the types in package.dhall needed to take on. I changed those and
it just worked.

The public API didn't change. This is quite cool.