Clean up the types
This moves the type to Haskell instead of deriving them from the
definitions in Dhall.

 - I need to write the serialization boilerplate myself before I rely on
   whatever was going on before.
 - I want more control over the field names, and I don't want to figure
   out how makeBaseFunctor can be customized.

This also features some changes to the type names and fields:

 - Decouple the recursive structure (execution order) from the essential
   information required to start a unit.
 - Allow for more orderings: now we can actually express trees, parallel
   starts, and serial starts. The StartOrder type can also contain
   arbitrary nesting: this hopefully makes it easy for users to express
   common patterns.

The ordering isn't perfect yet. I would like to express directed acyclic
graphs as well, but I'm not sure how that is best expressed yet.