Render clearline ANSI codes
This ensures that we do not mess up our output when the programs that we
start put the clear line escape code. We can render these if we assume
the programs we start do not send cursor movement codes.

There might be a way to communicate that to subprocesses using termcap.